requirements.txt
django 
python-dotenv
psycopg2-binary
stripe
djangorestframework
django-cors-headers
djangorestframework-simplejwt
flask 
flask-cors
httpx
websockets
langchain
langchain-community
twilio
channels
daphne
openai
requests
